+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Add builder node adder
 - Add ScBuilderScopeAdder
 - Add a needsUpdate method and needsUpdate supplier to allow triggering a builder on the basis of previous events
+- Add findItem method which returns either a Scope or Node
 
 ### Changed